+++ src/app/core/handle/handle.resource-type.ts
@@ -8,3 +8,7 @@ import { ResourceType } from '../shared/resource-type';
  */
 
 export const HANDLE = new ResourceType('handle');
+export const SUCCESSFUL_RESPONSE_START_CHAR = '2';
+export const COMMUNITY = 'Community';
+export const COLLECTION = 'Collection';
+export const ITEM = 'Item';
