+++ src/app/handle-page/edit-handle-page/edit-handle-page.component.html
@@ -2,7 +2,9 @@
   <form #editHandle = "ngForm" (ngSubmit) = "onClickSubmit(editHandle.value)">
     <div class="form-group">
       <label for="handle">{{'handle-table.edit-handle.form-handle-input-text' | translate}}</label>
-      <input type="text" class="form-control" id="handle" name="handle" [(ngModel)]="handle" [placeholder]="'handle-table.edit-handle.form-handle-input-placeholder' | translate">
+      <input type="text" class="form-control" id="handle" name="handle"
+             [(ngModel)]="handle"
+             [placeholder]="'handle-table.edit-handle.form-handle-input-placeholder' | translate">
     </div>
     <div class="form-group" *ngIf="resourceType">
       <label for="resourceType">{{'handle-table.table.resource-type' | translate}}</label>
@@ -20,12 +22,18 @@
     </div>
     <div class="form-group" *ngIf="resourceType == null">
       <label for="url">{{'handle-table.edit-handle.form-url-input-text' | translate}}</label>
-      <input type="text" class="form-control" id="url" name="url" [(ngModel)]="url" [placeholder]="'handle-table.edit-handle.form-url-input-placeholder' | translate">
+      <input type="text" class="form-control" id="url" name="url"
+             [(ngModel)]="url"
+             [placeholder]="'handle-table.edit-handle.form-url-input-placeholder' | translate">
     </div>
     <div class="form-check">
       <input type="checkbox" class="form-check-input" id="archive" name="archive" [(ngModel)]="archive">
-      <label class="form-check-label" for="archive">{{'handle-table.edit-handle.form-archive-input-check' | translate}}</label>
+      <label class="form-check-label" for="archive">
+        {{'handle-table.edit-handle.form-archive-input-check' | translate}}
+      </label>
     </div>
-    <button type="submit" class="btn btn-primary mt-2">{{'handle-table.edit-handle.form-button-submit' | translate}}</button>
+    <button type="submit" class="btn btn-primary mt-2">
+      {{'handle-table.edit-handle.form-button-submit' | translate}}
+    </button>
   </form>
 </div>
