+++ src/app/handle-page/handle-page-routing-paths.ts
@@ -1,5 +1,6 @@
+/**
+ * The routing paths
+ */
 export const HANDLE_TABLE_NEW_HANDLE_PATH = 'new-handle';
-
 export const HANDLE_TABLE_EDIT_HANDLE_PATH = 'edit-handle';
-
 export const GLOBAL_ACTIONS_PATH = 'change-handle-prefix';
