+++ src/app/handle-page/handle-page.component.html
@@ -1,12 +1,4 @@
 <div class="container">
-<!--  <div class="card">-->
-<!--    <h2 class="card-header">{{ 'handle-page.title' | translate }}</h2>-->
-<!--    <div class="card-body">-->
-<!--      <ds-handle-table></ds-handle-table>-->
-<!--      <ds-handle-global-actions></ds-handle-global-actions>-->
-<!--    </div>-->
-<!--  </div>-->
-
   <div class="row">
     <div class="col-12 pb-4">
       <h2 id="header" class="border-bottom pb-2">{{ 'handle-page.title' | translate }}</h2>
@@ -15,4 +7,3 @@
   <ds-handle-table></ds-handle-table>
   <ds-handle-global-actions></ds-handle-global-actions>
 </div>
-
