+++ src/app/handle-page/handle-page.component.ts
@@ -1,6 +1,9 @@
-import {ChangeDetectorRef, Component, OnInit} from '@angular/core';
-import {EndUserAgreementService} from '../core/end-user-agreement/end-user-agreement.service';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { EndUserAgreementService } from '../core/end-user-agreement/end-user-agreement.service';
 
+/**
+ * The component which contains the handle-table and the change-global-prefix section.
+ */
 @Component({
   selector: 'ds-handle-page',
   templateUrl: './handle-page.component.html',
@@ -10,12 +13,12 @@ export class HandlePageComponent implements OnInit {
 
   constructor(protected endUserAgreementService: EndUserAgreementService,
               private cdr: ChangeDetectorRef) {
-
   }
 
   /**
    * Initialize the component
    */
+  // tslint:disable-next-line:no-empty
   ngOnInit(): void {
   }
