+++ src/app/handle-page/handle-page.module.ts
@@ -1,18 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
 import { SharedModule } from '../shared/shared.module';
-import { FormModule } from '../shared/form/form.module';
-import {HandlePageComponent} from './handle-page.component';
-import {HandlePageRoutingModule} from './handle-page.routing.module';
-import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
-import {TranslateModule} from '@ngx-translate/core';
+import { HandlePageComponent } from './handle-page.component';
+import { HandlePageRoutingModule } from './handle-page.routing.module';
+import { TranslateModule } from '@ngx-translate/core';
 import { HandleTableComponent } from './handle-table/handle-table.component';
 import { HandleGlobalActionsComponent } from './handle-global-actions/handle-global-actions.component';
 import { NewHandlePageComponent } from './new-handle-page/new-handle-page.component';
 import { EditHandlePageComponent } from './edit-handle-page/edit-handle-page.component';
 import { ChangeHandlePrefixPageComponent } from './change-handle-prefix-page/change-handle-prefix-page.component';
-import {ReactiveFormsModule} from '@angular/forms';
+import { ReactiveFormsModule } from '@angular/forms';
 
 @NgModule({
   imports: [
