+++ src/app/handle-page/handle-page.routing.module.ts
@@ -1,25 +1,15 @@
 import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
-
 import { I18nBreadcrumbResolver } from '../core/breadcrumbs/i18n-breadcrumb.resolver';
-import {
-  SiteAdministratorGuard
-} from '../core/data/feature-authorization/feature-authorization-guard/site-administrator.guard';
-import { HandlePageModule } from './handle-page.module';
-import {HandlePageComponent} from './handle-page.component';
-import {COMMUNITY_EDIT_PATH} from '../community-page/community-page-routing-paths';
-import {CommunityPageAdministratorGuard} from '../community-page/community-page-administrator.guard';
+import { HandlePageComponent } from './handle-page.component';
 import {
   GLOBAL_ACTIONS_PATH,
   HANDLE_TABLE_EDIT_HANDLE_PATH,
   HANDLE_TABLE_NEW_HANDLE_PATH
 } from './handle-page-routing-paths';
-import {DeleteCommunityPageComponent} from '../community-page/delete-community-page/delete-community-page.component';
-import {AuthenticatedGuard} from '../core/auth/authenticated.guard';
-import {NewHandlePageComponent} from './new-handle-page/new-handle-page.component';
-import {EditHandlePageComponent} from './edit-handle-page/edit-handle-page.component';
-import {HandleGlobalActionsComponent} from './handle-global-actions/handle-global-actions.component';
-import {ChangeHandlePrefixPageComponent} from './change-handle-prefix-page/change-handle-prefix-page.component';
+import { NewHandlePageComponent } from './new-handle-page/new-handle-page.component';
+import { EditHandlePageComponent } from './edit-handle-page/edit-handle-page.component';
+import { ChangeHandlePrefixPageComponent } from './change-handle-prefix-page/change-handle-prefix-page.component';
 
 @NgModule({
   imports: [
