+++ src/app/handle-page/handle-table/handle-table.component.spec.ts
@@ -1,24 +1,24 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { HandleTableComponent } from './handle-table.component';
-import {HandleDataService} from '../../core/data/handle-data.service';
-import {PaginationService} from '../../core/pagination/pagination.service';
-import {Router} from '@angular/router';
-import {RequestService} from '../../core/data/request.service';
-import {createSuccessfulRemoteDataObject$} from '../../shared/remote-data.utils';
-import {of as observableOf} from 'rxjs';
-import {SharedModule} from '../../shared/shared.module';
-import {CommonModule} from '@angular/common';
-import {ReactiveFormsModule} from '@angular/forms';
-import {TranslateModule} from '@ngx-translate/core';
-import {RouterTestingModule} from '@angular/router/testing';
-import {Handle} from '../../core/handle/handle.model';
-import {PaginationServiceStub} from '../../shared/testing/pagination-service.stub';
-import {RouterStub} from '../../shared/testing/router.stub';
-import {getHandleTableModulePath} from '../../app-routing-paths';
-import {defaultPagination} from './handle-table-pagination';
-import {buildPaginatedList} from '../../core/data/paginated-list.model';
-import {PageInfo} from '../../core/shared/page-info.model';
-import {HANDLE_TABLE_EDIT_HANDLE_PATH} from '../handle-page-routing-paths';
+import { HandleDataService } from '../../core/data/handle-data.service';
+import { PaginationService } from '../../core/pagination/pagination.service';
+import { Router } from '@angular/router';
+import { RequestService } from '../../core/data/request.service';
+import { createSuccessfulRemoteDataObject$ } from '../../shared/remote-data.utils';
+import { of as observableOf } from 'rxjs';
+import { SharedModule } from '../../shared/shared.module';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+import { TranslateModule } from '@ngx-translate/core';
+import { RouterTestingModule } from '@angular/router/testing';
+import { Handle } from '../../core/handle/handle.model';
+import { PaginationServiceStub } from '../../shared/testing/pagination-service.stub';
+import { RouterStub } from '../../shared/testing/router.stub';
+import { getHandleTableModulePath } from '../../app-routing-paths';
+import { defaultPagination } from './handle-table-pagination';
+import { buildPaginatedList } from '../../core/data/paginated-list.model';
+import { PageInfo } from '../../core/shared/page-info.model';
+import { HANDLE_TABLE_EDIT_HANDLE_PATH } from '../handle-page-routing-paths';
 
 /**
  * The test for testing HandleTableComponent.
