+++ src/app/handle-page/handle-table/handle-table.component.ts
@@ -1,61 +1,35 @@
-import {ChangeDetectorRef, Component, ElementRef, OnInit, ViewChild} from '@angular/core';
-import {BehaviorSubject, combineLatest as observableCombineLatest, combineLatest, fromEvent, Observable} from 'rxjs';
-import {RemoteData} from '../../core/data/remote-data';
-import {PaginatedList} from '../../core/data/paginated-list.model';
-import {Version} from '../../core/shared/version.model';
-import {COLLECTION, COMMUNITY, Handle, ITEM, SUCCESSFUL_RESPONSE_START_CHAR} from '../../core/handle/handle.model';
-import {HandleDataService} from '../../core/data/handle-data.service';
-import {PaginationComponentOptions} from '../../shared/pagination/pagination-component-options.model';
-import {
-  debounceTime,
-  distinctUntilChanged,
-  distinctUntilKeyChanged,
-  first,
-  last,
-  map, startWith,
-  switchMap, take,
-  takeLast,
-  takeUntil, tap
-} from 'rxjs/operators';
-import {VersionHistory} from '../../core/shared/version-history.model';
-import {PaginatedSearchOptions} from '../../shared/search/models/paginated-search-options.model';
-import {followLink} from '../../shared/utils/follow-link-config.model';
-import {
-  getAllCompletedRemoteData,
-  getFirstCompletedRemoteData, getFirstSucceededRemoteData,
-  getFirstSucceededRemoteDataPayload, getPaginatedListPayload, getRemoteDataPayload, toDSpaceObjectListRD
-} from '../../core/shared/operators';
-import {PaginationService} from '../../core/pagination/pagination.service';
-import {ObjectSelectComponent} from '../../shared/object-select/object-select/object-select.component';
-import {Collection} from '../../core/shared/collection.model';
-import {ObjectSelectService} from '../../shared/object-select/object-select.service';
-import {AuthorizationDataService} from '../../core/data/feature-authorization/authorization-data.service';
-import {getHandleTableModulePath, getPageNotFoundRoute} from '../../app-routing-paths';
-import {HANDLE_TABLE_EDIT_HANDLE_PATH, HANDLE_TABLE_NEW_HANDLE_PATH} from '../handle-page-routing-paths';
-import {isEmpty, isNotEmpty} from '../../shared/empty.util';
-import {Router} from '@angular/router';
-import {Operation} from 'fast-json-patch';
-import {DeleteRequest, PatchRequest} from '../../core/data/request.models';
-import {RequestService} from '../../core/data/request.service';
-import wait from 'fork-ts-checker-webpack-plugin/lib/utils/async/wait';
-import {
-  defaultPagination,
-  defaultSortConfiguration,
-  paginationID,
-  redirectBackWithPaginationOption
-} from './handle-table-pagination';
-import {TranslateService} from '@ngx-translate/core';
-import {NotificationsService} from '../../shared/notifications/notifications.service';
-import {DSpaceObjectType} from '../../core/shared/dspace-object-type.model';
-import {Item} from '../../core/shared/item.model';
-import {SortOptions} from '../../core/cache/models/sort-options.model';
+import { ChangeDetectorRef, Component, ElementRef, OnInit, ViewChild } from '@angular/core';
+import { BehaviorSubject, combineLatest as observableCombineLatest, fromEvent } from 'rxjs';
+import { RemoteData } from '../../core/data/remote-data';
+import { PaginatedList } from '../../core/data/paginated-list.model';
+import { HandleDataService } from '../../core/data/handle-data.service';
+import { PaginationComponentOptions } from '../../shared/pagination/pagination-component-options.model';
+import { debounceTime, distinctUntilChanged, switchMap, take } from 'rxjs/operators';
+import { getFirstSucceededRemoteData, getRemoteDataPayload } from '../../core/shared/operators';
+import { PaginationService } from '../../core/pagination/pagination.service';
+import { getHandleTableModulePath } from '../../app-routing-paths';
+import { HANDLE_TABLE_EDIT_HANDLE_PATH, HANDLE_TABLE_NEW_HANDLE_PATH } from '../handle-page-routing-paths';
+import { isEmpty, isNotEmpty } from '../../shared/empty.util';
+import { Router } from '@angular/router';
+import { DeleteRequest } from '../../core/data/request.models';
+import { RequestService } from '../../core/data/request.service';
+import { defaultPagination, defaultSortConfiguration } from './handle-table-pagination';
+import { TranslateService } from '@ngx-translate/core';
+import { NotificationsService } from '../../shared/notifications/notifications.service';
+import { SortOptions } from '../../core/cache/models/sort-options.model';
+import { Handle } from '../../core/handle/handle.model';
+import { COLLECTION, COMMUNITY, ITEM, SUCCESSFUL_RESPONSE_START_CHAR } from '../../core/handle/handle.resource-type';
 
+/**
+ * Constants for converting the searchQuery for the server
+ */
 export const HANDLE_SEARCH_OPTION = 'handle';
-export const INTERNAL_SEARCH_OPTION = 'internal';
 export const URL_SEARCH_OPTION = 'url';
 export const RESOURCE_TYPE_SEARCH_OPTION = 'resourceTypeId';
 
-
+/**
+ * The component which contains the Handle table and search panel for filtering the handles.
+ */
 @Component({
   selector: 'ds-handle-table',
   templateUrl: './handle-table.component.html',
@@ -72,6 +46,9 @@ export class HandleTableComponent implements OnInit {
               private notificationsService: NotificationsService,) {
   }
 
+  /**
+   * The reference for the input html element
+   */
   @ViewChild('searchInput', {static: true}) searchInput: ElementRef;
 
   /**
@@ -90,35 +67,65 @@ export class HandleTableComponent implements OnInit {
    */
   options: PaginationComponentOptions;
 
+  /**
+   * The configuration which is send to the server with search request.
+   */
   sortConfiguration: SortOptions;
 
+  /**
+   * The value typed in the search panel.
+   */
   searchQuery = '';
 
+  /**
+   * Filter the handles based on this column.
+   */
   searchOption: string;
 
   /**
-   * @TODO docs
+   * String value of the `Handle` search option. This value is loaded from the `en.json5`.
+   */
+  handleOption: string;
+
+  /**
+   * String value of the `Internal` search option. This value is loaded from the `en.json5`.
+   */
+  internalOption: string;
+
+  /**
+   * String value of the `Resource type` search option. This value is loaded from the `en.json5`.
+   */
+  resourceTypeOption: string;
+
+  /**
+   * If the request isn't processed show to loading bar.
    */
   isLoading = false;
 
+  /**
+   * The handle redirection link.
+   */
   handleRoute: string;
 
+  /**
+   * The new handle redirection link.
+   */
   newHandlePath = HANDLE_TABLE_NEW_HANDLE_PATH;
 
+  /**
+   * The edit handle redirection link.
+   */
   editHandlePath = HANDLE_TABLE_EDIT_HANDLE_PATH;
 
+  /**
+   * The handle which is selected in the handle table.
+   */
   selectedHandle = null;
-
-  handleOption: string;
-
-  internalOption: string;
-
-  resourceTypeOption: string;
 
   ngOnInit(): void {
     this.handleRoute = getHandleTableModulePath();
     this.initializePaginationOptions();
-    this.sortConfiguration = defaultSortConfiguration;
+    this.initializeSortingOptions();
     this.getAllHandles();
 
     this.handleOption = this.translateService.instant('handle-table.table.handle');
@@ -126,14 +133,14 @@ export class HandleTableComponent implements OnInit {
     this.resourceTypeOption = this.translateService.instant('handle-table.table.resource-type');
   }
 
-  private initializePaginationOptions() {
-    this.options = defaultPagination;
-  }
-
+  /**
+   * Load all handles based on the pagination and sorting options.
+   */
   getAllHandles() {
     this.handlesRD$ = new BehaviorSubject<RemoteData<PaginatedList<Handle>>>(null);
     this.isLoading = true;
 
+    // load the current pagination and sorting options
     const currentPagination$ = this.paginationService.getCurrentPagination(this.options.id, this.options);
     const currentSort$ = this.paginationService.getCurrentSort(this.options.id, this.sortConfiguration);
 
@@ -160,6 +167,10 @@ export class HandleTableComponent implements OnInit {
     this.getAllHandles();
   }
 
+  /**
+   * Mark the handle as selected or unselect if it is already clicked.
+   * @param handleId id of the selected handle
+   */
   switchSelectedHandle(handleId) {
     if (this.selectedHandle === handleId) {
       this.selectedHandle = null;
@@ -168,12 +179,18 @@ export class HandleTableComponent implements OnInit {
     }
   }
 
+  /**
+   * Redirect to the new handle component with the current pagination options.
+   */
   redirectWithCurrentPage() {
     this.router.navigate([this.handleRoute, this.newHandlePath],
       { queryParams: { currentPage: this.options.currentPage } },
     );
   }
 
+  /**
+   * Redirect to the edit handle component with the handle attributes passed in the url.
+   */
   redirectWithHandleParams() {
     // check if is selected some handle
     if (isEmpty(this.selectedHandle)) {
@@ -198,6 +215,9 @@ export class HandleTableComponent implements OnInit {
     });
   }
 
+  /**
+   * Delete selected handle
+   */
   deleteHandles() {
     // check if is selected some handle
     if (isEmpty(this.selectedHandle)) {
@@ -249,6 +269,11 @@ export class HandleTableComponent implements OnInit {
       });
   }
 
+  /**
+   * Deleted handle must be removed from the table. Wait for removing the handle from the server and then load
+   * the handles again.
+   * @param deletedHandleId
+   */
   public refreshTableAfterDelete(deletedHandleId) {
     let counter = 0;
     // The timeout for checking if the handle was daleted in the database
@@ -291,6 +316,9 @@ export class HandleTableComponent implements OnInit {
     }, 250 );
   }
 
+  /**
+   * If the user is typing the searchQuery is changing.
+   */
   setSearchQuery() {
     if (isEmpty(this.searchOption)) {
       return;
@@ -306,12 +334,20 @@ export class HandleTableComponent implements OnInit {
       });
   }
 
+  /**
+   * The search option is selected from the dropdown menu.
+   * @param event with the selected value
+   */
   setSearchOption(event) {
     this.searchOption = event?.target?.innerHTML;
     this.searchInput.nativeElement.value = '';
     this.searchHandles('');
   }
 
+  /**
+   * Update the sortConfiguration based on the `searchOption` and the `searchQuery` but parse that attributes at first.
+   * @param searchQuery
+   */
   searchHandles(searchQuery = '') {
     if (isEmpty(this.searchOption)) {
       return;
@@ -358,4 +394,11 @@ export class HandleTableComponent implements OnInit {
     this.getAllHandles();
   }
 
+  private initializePaginationOptions() {
+    this.options = defaultPagination;
+  }
+
+  private initializeSortingOptions() {
+    this.sortConfiguration = defaultSortConfiguration;
+  }
 }
