+++ src/assets/i18n/en.json5
@@ -1644,6 +1644,9 @@
 
   "handle-table.button.delete-handle": "Delete handle",
 
+  "handle-table.dropdown.search-option": "Search option",
+
+
 
   "handle-table.global-actions.title": "Global Actions",
